% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{darken_colour}
\alias{darken_colour}
\alias{lighten_colour}
\title{Darken/lighten a hex colour by the given amount
Stolen from \url{https://benjaminlmoore.wordpress.com/2014/03/18/guardian-data-blog-uk-elections/}}
\usage{
darken_colour(hex_colour, amount = 0.15)

lighten_colour(hex_colour, amount = 0.15)
}
\arguments{
\item{hex_colour}{strings e.g. "#345678"}

\item{amount}{fraction to darken by. default 0.15}
}
\value{
darkened hex colours
}
\description{
Darken/lighten a hex colour by the given amount
Stolen from \url{https://benjaminlmoore.wordpress.com/2014/03/18/guardian-data-blog-uk-elections/}
}
