% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-path.R
\name{geom_path_3d}
\alias{geom_path_3d}
\alias{geom_line_3d}
\title{An extension of \code{geom_path} and \code{geom_line} into the third dimension.}
\usage{
geom_path_3d(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  extrude = FALSE,
  material = list(),
  keep2d = FALSE
)

geom_line_3d(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  extrude = FALSE,
  material = list(),
  keep2d = FALSE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, ..., lineend, linejoin, linemitre, arrow, na.rm, show.legend, inherit.aes, orientation}{See documentation for \code{ggplot2::geom_path}}

\item{extrude}{whether or not to extrude the polygon}

\item{material}{Arguments passed to \code{rgl::material3d()} to specify the
material properties for this geom.   Any
parameters specified in this argument override the defaults.  See \code{rgl}
documentation for more info. Also see \code{ggrgl::standard_material} for
the standard default material parameters used for all objects.}

\item{keep2d}{keep the original 2d representation? default: FALSE}
}
\description{
An extension of \code{geom_path} and \code{geom_line} into the third dimension.
}
\section{Full 3d Positioning}{

The \code{3d} suffix indicates that this geom allows for the full 3d
positioning of each node and/or edge which makes it up.  Thus the resulting
geometrical element rendered on the plot can appear in any orientiation, and
is not limited to planar representation.
}

\section{Geom Supports Extrusion}{

This geometry supports extrusion.  Set \code{extrude = TRUE} and adjust
aesthetics:
\describe{
\item{\code{extrude_face_fill}}{The colour of the extruded faces. Default: 'grey20'}
\item{\code{extrude_face_alpha}}{Alpha for the extruded faces. Default: 1}
\item{\code{extrude_edge_colour}}{Colour of the edges of the extrusion. Default: NA (invisible)}
\item{\code{extrude_edge_alpha}}{Alpha for the extruded edges. Default: 1}
\item{\code{extrude_edge_size}}{Width of the line to draw the extruded edge. Default: 1}
\item{\code{extrude_z}}{The lower limit of the extrudsion. Default: 0.05}
}
}

\seealso{
Other 3d geoms: 
\code{\link{geom_segment_3d}()},
\code{\link{geom_sphere_3d}()}

Other geoms supporting extrusion: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_point_z}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_ribbon_z}()},
\code{\link{geom_segment_3d}()},
\code{\link{geom_tile_z}()}
}
\concept{3d geoms}
\concept{geoms supporting extrusion}
