% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point.R
\name{geom_point_z}
\alias{geom_point_z}
\title{An extension of \code{geom_point} into the third dimension.}
\usage{
geom_point_z(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  extrude = FALSE,
  material = list(),
  npolys = 10,
  keep2d = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
documentation for \code{ggplot2::geom_point()}}

\item{extrude}{whether or not to extrude the polygon}

\item{material}{Arguments passed to \code{rgl::material3d()} to specify the
material properties for this geom.   Any
parameters specified in this argument override the defaults.  See \code{rgl}
documentation for more info. Also see \code{ggrgl::standard_material} for
the standard default material parameters used for all objects.}

\item{npolys}{Circles in ggrgl are approximated by polygons. \code{npolys}
controls how many polygons are used to represent each cicle. Default: 10}

\item{keep2d}{keep the original 2d representation? default: FALSE}
}
\description{
Note that points in \code{ggplot2} are circles,  and this function still
plots circles, but with the option of raising or lowering the circle in the
\code{z} direction.
}
\details{
For points with a fully 3-dimensional appearance see \code{geom_sphere_3d}.
}
\section{Z Offset Geom}{

The \code{z} suffix indicates that this geom produces planar graphical elements,
parallel to the ground i.e. the drawn element is the same as that for
\code{ggplot2} except it has a vertical offset (constant across each graphical element).
}

\section{Geom Supports Extrusion}{

This geometry supports extrusion.  Set \code{extrude = TRUE} and adjust
aesthetics:
\describe{
\item{\code{extrude_face_fill}}{The colour of the extruded faces. Default: 'grey20'}
\item{\code{extrude_face_alpha}}{Alpha for the extruded faces. Default: 1}
\item{\code{extrude_edge_colour}}{Colour of the edges of the extrusion. Default: NA (invisible)}
\item{\code{extrude_edge_alpha}}{Alpha for the extruded edges. Default: 1}
\item{\code{extrude_edge_size}}{Width of the line to draw the extruded edge. Default: 1}
\item{\code{extrude_z}}{The lower limit of the extrudsion. Default: 0.05}
}
}

\seealso{
Other zoffset geoms: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_ribbon_z}()},
\code{\link{geom_text_z}()},
\code{\link{geom_tile_z}()}

Other geoms supporting extrusion: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_path_3d}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_ribbon_z}()},
\code{\link{geom_segment_3d}()},
\code{\link{geom_tile_z}()}
}
\concept{geoms supporting extrusion}
\concept{zoffset geoms}
