% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ribbon.R
\name{geom_ribbon_z}
\alias{geom_ribbon_z}
\alias{geom_area_z}
\title{An extension of \code{geom_ribbon} and \code{geom_area} into the third dimension.}
\usage{
geom_ribbon_z(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "both",
  extrude = FALSE,
  material = list(),
  keep2d = FALSE
)

geom_area_z(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "stack",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  outline.type = "upper",
  extrude = FALSE,
  material = list(),
  keep2d = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, orientation, show.legend, inherit.aes, outline.type}{See documentation for \code{ggplot2::geom_ribbon}}

\item{extrude}{whether or not to extrude the polygon}

\item{material}{Arguments passed to \code{rgl::material3d()} to specify the
material properties for this geom.   Any
parameters specified in this argument override the defaults.  See \code{rgl}
documentation for more info. Also see \code{ggrgl::standard_material} for
the standard default material parameters used for all objects.}

\item{keep2d}{keep the original 2d representation? default: FALSE}
}
\description{
An extension of \code{geom_ribbon} and \code{geom_area} into the third dimension.
}
\section{Z Offset Geom}{

The \code{z} suffix indicates that this geom produces planar graphical elements,
parallel to the ground i.e. the drawn element is the same as that for
\code{ggplot2} except it has a vertical offset (constant across each graphical element).
}

\section{Geom Supports Extrusion}{

This geometry supports extrusion.  Set \code{extrude = TRUE} and adjust
aesthetics:
\describe{
\item{\code{extrude_face_fill}}{The colour of the extruded faces. Default: 'grey20'}
\item{\code{extrude_face_alpha}}{Alpha for the extruded faces. Default: 1}
\item{\code{extrude_edge_colour}}{Colour of the edges of the extrusion. Default: NA (invisible)}
\item{\code{extrude_edge_alpha}}{Alpha for the extruded edges. Default: 1}
\item{\code{extrude_edge_size}}{Width of the line to draw the extruded edge. Default: 1}
\item{\code{extrude_z}}{The lower limit of the extrudsion. Default: 0.05}
}
}

\seealso{
Other zoffset geoms: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_point_z}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_text_z}()},
\code{\link{geom_tile_z}()}

Other geoms supporting extrusion: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_path_3d}()},
\code{\link{geom_point_z}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_segment_3d}()},
\code{\link{geom_tile_z}()}
}
\concept{geoms supporting extrusion}
\concept{zoffset geoms}
