% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text.R
\name{geom_text_z}
\alias{geom_text_z}
\title{An extension of \code{geom_text} into the third dimension.}
\usage{
geom_text_z(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  material = list(),
  keep2d = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, ..., parse, nudge_x, nudge_y, check_overlap, na.rm, show.legend, inherit.aes}{See documentation for \code{ggplot::geom_text}}

\item{material}{Arguments passed to \code{rgl::material3d()} to specify the
material properties for this geom.   Any
parameters specified in this argument override the defaults.  See \code{rgl}
documentation for more info. Also see \code{ggrgl::standard_material} for
the standard default material parameters used for all objects.}

\item{keep2d}{keep the original 2d representation? default: FALSE}
}
\description{
An extension of \code{geom_text} into the third dimension.
}
\section{Z Offset Geom}{

The \code{z} suffix indicates that this geom produces planar graphical elements,
parallel to the ground i.e. the drawn element is the same as that for
\code{ggplot2} except it has a vertical offset (constant across each graphical element).
}

\seealso{
Other zoffset geoms: 
\code{\link{geom_bar_z}()},
\code{\link{geom_contour_filled_z}()},
\code{\link{geom_density_z}()},
\code{\link{geom_point_z}()},
\code{\link{geom_polygon_z}()},
\code{\link{geom_rect_z}()},
\code{\link{geom_ribbon_z}()},
\code{\link{geom_tile_z}()}
}
\concept{zoffset geoms}
