% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-rect.R, R/geom-bar.R, R/geom-polygon.R,
%   R/geom-contour-filled.R, R/geom-path.R, R/geom-contour.R, R/geom-ribbon.R,
%   R/geom-density.R, R/geom-point.R, R/geom-segment.R, R/geom-sphere.R,
%   R/geom-text.R, R/geom-tile.R, R/stat-contour-filled.R, R/stat-contour.R
\docType{data}
\name{GeomRectZ}
\alias{GeomRectZ}
\alias{GeomBarZ}
\alias{GeomPolygonZ}
\alias{GeomContourFilledZ}
\alias{GeomPath3d}
\alias{GeomLine3d}
\alias{GeomContourZ}
\alias{GeomRibbonZ}
\alias{GeomAreaZ}
\alias{GeomDensityZ}
\alias{GeomPointZ}
\alias{GeomSegment3d}
\alias{GeomSphere3d}
\alias{GeomTextZ}
\alias{GeomTileZ}
\alias{StatContourFilledZ}
\alias{StatContourZ}
\title{GeomRectZ}
\description{
GeomRectZ

GeomBarZ

GeomPolygonZ

GeomPath3d

GeomLine3d

GeomContourZ

GeomRibbonZ

GeomRibbonZ

GeomDensitZ

GeomPointZ

GeomSegment3d

GeomSphere3d

GeomTextZ

GeomTileZ
}
\keyword{datasets}
