% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-extrude-hue.R
\name{scale_extrude_edge_colour_hue}
\alias{scale_extrude_edge_colour_hue}
\alias{scale_extrude_face_fill_hue}
\alias{scale_extrude_face_fill_discrete}
\alias{scale_extrude_edge_colour_discrete}
\title{Evenly spaced colours for discrete data}
\usage{
scale_extrude_edge_colour_hue(
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50",
  aesthetics = "extrude_edge_colour"
)

scale_extrude_face_fill_hue(
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50",
  aesthetics = "extrude_face_fill"
)
}
\arguments{
\item{h, c, l, h.start, direction, ...}{See \code{ggplot2::scale_colour_hue}}

\item{na.value}{Colour to use for missing values}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the `colour` and `fill` aesthetics at the
same time, via `aesthetics = c("colour", "fill")`.}
}
\description{
This is the default colour scale for categorical variables. It maps each
level to an evenly spaced hue on the colour wheel. It does not generate
colour-blind safe palettes.
}
