% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-extrude-colour.R
\name{scale_extrude_face_fill_continuous}
\alias{scale_extrude_face_fill_continuous}
\title{Continuous and binned colour scales}
\usage{
scale_extrude_face_fill_continuous(
  ...,
  type = getOption("ggplot2.continuous.fill", default = "gradient")
)
}
\arguments{
\item{...}{Additional parameters passed on to the scale type}

\item{type}{One of "gradient" (the default) or "viridis" indicating the
colour scale to use}
}
\description{
See \code{ggplot2::scale_colour_continuous()} for more information
}
