% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-extrude-gradient.R
\name{scale_extrude_face_fill_gradient}
\alias{scale_extrude_face_fill_gradient}
\alias{scale_extrude_face_fill_gradient2}
\alias{scale_extrude_face_fill_gradientn}
\title{Gradient colour scales}
\usage{
scale_extrude_face_fill_gradient(
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = FALSE,
  aesthetics = "extrude_face_fill"
)

scale_extrude_face_fill_gradient2(
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "extrude_face_fill"
)

scale_extrude_face_fill_gradientn(
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "extrude_face_fill",
  colors
)

scale_extrude_face_fill_gradientn(
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "extrude_face_fill",
  colors
)
}
\arguments{
\item{low, mid, high, midpoint}{Colours for low and high ends of the gradient.}

\item{space, ..., na.value, aesthetics}{See \code{scales::seq_gradient_pal}, \code{scale_colour_hue}, \code{ggplot2::continuous_scale}}

\item{guide}{Type of legend. Use `"colourbar"` for continuous
colour bar, or `"legend"` for discrete colour legend.}

\item{colours, colors}{Vector of colours to use for n-colour gradient.}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link[scales:rescale]{rescale()}} for a convenience function
to map an arbitrary range to between 0 and 1.}
}
\description{
See \code{ggplot2::scale_colour_gradient()} for more information
}
\details{
`scale_*_gradient` creates a two colour gradient (low-high),
`scale_*_gradient2` creates a diverging colour gradient (low-mid-high),
`scale_*_gradientn` creates a n-colour gradient.
}
