% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-identity.R
\name{scale_z_identity}
\alias{scale_z_identity}
\title{Use z values without scaling}
\usage{
scale_z_identity(..., guide = "none")
}
\arguments{
\item{..., guide}{See \code{ggplot2} for documentation on identity scales.
e.g. \code{ggplot2::scale_alpha_identity()}}
}
\description{
Use z values without scaling
}
