% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-zend.R
\name{scale_zend_continuous}
\alias{scale_zend_continuous}
\alias{scale_zend_discrete}
\alias{scale_zend_ordinal}
\title{Scales for zend}
\usage{
scale_zend_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 100),
  trans = "identity",
  guide = FALSE
)
}
\arguments{
\item{name, breaks, labels, limits, trans, guide}{See \code{ggplot2::scale_size} for more information}

\item{range}{a numeric vector of length 2 that specifies the minimum and
maximum size of the plotting symbol after transformation.}
}
\description{
Scales for zend
}
