% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggrgl}
\alias{theme_ggrgl}
\title{A theme based upon \code{theme_bw()} which sets some good defaults for \code{ggrgl} plots}
\usage{
theme_ggrgl(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size, base_family, base_line_size, base_rect_size}{See docs for
\code{ggplot2::theme_bw()}}
}
\description{
A theme based upon \code{theme_bw()} which sets some good defaults for \code{ggrgl} plots
}
