% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volcano_arr}
\alias{volcano_arr}
\title{Volcano dataset as an RGB array coloured with `viridis` palette}
\format{
An object of class \code{array} of dimension 61 x 87 x 3.
}
\usage{
volcano_arr
}
\description{
Volcano dataset as an RGB array coloured with `viridis` palette
}
\keyword{datasets}
